Add regexpFilter extension property
This property when set will take precedence on the existing filter one.
It allows for one to provide a regular expression that will be used to
filter the file paths to include when weaving the files in the project.

This can be useful when one want to weave the code in the project as
well as specific libraries.